+++ tests/standalone/io/https_bad_certificate_test.dart
@@ -12,12 +12,11 @@ import "package:expect/expect.dart";
 final HOST_NAME = 'localhost';
 
 String localFile(path) => Platform.script.resolve(path).toFilePath();
-List<int> readLocalFile(path) => (new File(localFile(path))).readAsBytesSync();
 
 SecurityContext serverContext = new SecurityContext()
-  ..useCertificateChainBytes(readLocalFile('certificates/server_chain.pem'))
-  ..usePrivateKeyBytes(readLocalFile('certificates/server_key.pem'),
-                         password: 'dartdart');
+  ..useCertificateChainSync(localFile('certificates/server_chain.pem'))
+  ..usePrivateKeySync(localFile('certificates/server_key.pem'),
+                      password: 'dartdart');
 
 class CustomException {}
 
@@ -32,7 +31,7 @@ main() async {
   });
 
   SecurityContext goodContext = new SecurityContext()
-    ..setTrustedCertificates(file: localFile('certificates/trusted_certs.pem'));
+    ..setTrustedCertificatesSync(localFile('certificates/trusted_certs.pem'));
   SecurityContext badContext = new SecurityContext();
   SecurityContext defaultContext = SecurityContext.defaultContext;