+++ tests/standalone/io/https_unauthorized_test.dart
@@ -14,16 +14,15 @@ const HOST_NAME = "localhost";
 const CERTIFICATE = "localhost_cert";
 
 String localFile(path) => Platform.script.resolve(path).toFilePath();
-List<int> readLocalFile(path) => (new File(localFile(path))).readAsBytesSync();
 
 SecurityContext untrustedServerContext = new SecurityContext()
-  ..useCertificateChainBytes(readLocalFile(
+  ..useCertificateChainSync(localFile(
       'certificates/untrusted_server_chain.pem'))
-  ..usePrivateKeyBytes(readLocalFile('certificates/untrusted_server_key.pem'),
-                         password: 'dartdart');
+  ..usePrivateKeySync(localFile('certificates/untrusted_server_key.pem'),
+                      password: 'dartdart');
 
 SecurityContext clientContext = new SecurityContext()
-  ..setTrustedCertificates(file: localFile('certificates/trusted_certs.pem'));
+  ..setTrustedCertificatesSync(localFile('certificates/trusted_certs.pem'));
 
 Future<SecureServerSocket> runServer() {
   return HttpServer.bindSecure(