+++ tests/standalone/io/raw_secure_server_socket_test.dart
@@ -15,15 +15,14 @@ import "package:expect/expect.dart";
 
 InternetAddress HOST;
 String localFile(path) => Platform.script.resolve(path).toFilePath();
-List<int> readLocalFile(path) => (new File(localFile(path))).readAsBytesSync();
 
 SecurityContext serverContext = new SecurityContext()
-  ..useCertificateChainBytes(readLocalFile('certificates/server_chain.pem'))
-  ..usePrivateKeyBytes(readLocalFile('certificates/server_key.pem'),
-                       password: 'dartdart');
+  ..useCertificateChainSync(localFile('certificates/server_chain.pem'))
+  ..usePrivateKeySync(localFile('certificates/server_key.pem'),
+                      password: 'dartdart');
 
 SecurityContext clientContext = new SecurityContext()
-  ..setTrustedCertificates(file: localFile('certificates/trusted_certs.pem'));
+  ..setTrustedCertificatesSync(localFile('certificates/trusted_certs.pem'));
 
 void testSimpleBind() {
   asyncStart();
@@ -575,13 +574,13 @@ runTests() {
   var chain =
       Platform.script.resolve('certificates/untrusted_server_chain.pem')
       .toFilePath();
-  context.useCertificateChain(chain);
+  context.useCertificateChainSync(chain);
   testSimpleConnectFail(context, false);
   testSimpleConnectFail(context, true);
   var key =
       Platform.script.resolve('certificates/untrusted_server_key.pem')
        .toFilePath();
-  context.usePrivateKey(key, password: 'dartdart');
+  context.usePrivateKeySync(key, password: 'dartdart');
   testSimpleConnectFail(context, false);
   testSimpleConnectFail(context, true);
   testServerListenAfterConnect();